Update DeliveryAddressSelector screen

+++ core/domain/src/main/java/ru/livetyping/zarina/core/domain/model/checkout/DeliveryOption.kt
@@ -1,11 +1,13 @@
 package ru.livetyping.zarina.core.domain.model.checkout
 
+import java.math.BigDecimal
+
 // Marked as stable on config/compose/stability_config.txt
 public data class DeliveryOption(
     val id: Id,
     val title: String,
     val description: String,
-    val price: Int,
+    val price: BigDecimal,
     val dateTimePeriods: List<DateTimePeriod>,
 ) {
     // Marked as stable on config/compose/stability_config.txt
